HACKFEST-VINNY: Refactor Workflow UI URL and directory structure
