+++ app/frontends/jobs.ts
@@ -1,11 +1,8 @@
 import { Response, NextFunction } from 'express';
-import { sanitizeImage } from 'app/util/string';
 import { Job, JobStatus, JobQuery } from '../models/job';
 import { keysToLowerCase } from '../util/object';
-import isUUID from '../util/uuid';
-import cancelAndSaveJob from '../util/job';
+import cancelAndSaveJob, { validateJobId } from '../util/job';
 import JobLink from '../models/job-link';
-import { getWorkItemsByJobId, WorkItemStatus } from '../models/work-item';
 import { needsStacLink } from '../util/stac';
 import { getRequestRoot } from '../util/url';
 import { getCloudAccessJsonLink, getCloudAccessShLink, getStacCatalogLink, getStatusLink, Link } from '../util/links';
@@ -137,16 +134,6 @@ export async function getJobsListing(
 }
 
 /**
- * Throws an exception if the JobID is not in the valid format for a jobID.
- * @param jobID - The jobID to validate
- */
-function validateJobId(jobID: string): void {
-  if (!isUUID(jobID)) {
-    throw new RequestValidationError(`Invalid format for Job ID '${jobID}'. Job ID must be a UUID.`);
-  }
-}
-
-/**
  * Express.js handler that returns job status for a single job `(/jobs/{jobID})`
  *
  * @param req - The request sent by the client
@@ -231,160 +218,5 @@ export async function cancelJob(
     } else {
       next(e);
     }
-  }
-}
-
-/**
- * Display jobs along with their status in the workflow UI.
- *
- * @param req - The request sent by the client
- * @param res - The response to send to the client
- * @param next - The next function in the call chain
- * @returns HTML page of clickable jobs which take the user to a
- * page where they can visualize the whole workflow as it happens
- */
-export async function getJobsForWorkflowUI(
-  req: HarmonyRequest, res: Response, next: NextFunction,
-): Promise<void> {
-  const badgeClasses = {};
-  badgeClasses[JobStatus.ACCEPTED] = 'primary';
-  badgeClasses[JobStatus.CANCELED] = 'secondary';
-  badgeClasses[JobStatus.FAILED] = 'danger';
-  badgeClasses[JobStatus.SUCCESSFUL] = 'success';
-  badgeClasses[JobStatus.RUNNING] = 'info';
-  try {
-    const { page, limit } = getPagingParams(req, env.defaultJobListPageSize);
-    const query: JobQuery = {};
-    if (!req.context.isAdminAccess) {
-      query.username = req.user;
-    }
-    const { data: jobs, pagination } = await Job.queryAll(db, query, false, page, limit);
-    const pageLinks = getPagingLinks(req, pagination);
-    const nextPage = pageLinks.find((l) => l.rel === 'next');
-    const previousPage = pageLinks.find((l) => l.rel === 'prev');
-    setPagingHeaders(res, pagination);
-    res.render('jobs/workflow-jobs', {
-      jobs,
-      jobBadge() { return badgeClasses[this.status]; },
-      jobUrl() {
-        const url = new URL(this.request);
-        const path = url.pathname + url.search;
-        return path;
-      },
-      links: [
-        { ...previousPage, linkTitle: 'previous' },
-        { ...nextPage, linkTitle: 'next' },
-      ],
-      linkDisabled() { return (this.href ? '' : 'disabled'); },
-      linkHref() { return (this.href || ''); },
-    });
-  } catch (e) {
-    req.context.logger.error(e);
-    next(e);
-  }
-}
-
-/**
- * Display a job's progress and work items in the workflow UI.
- *
- * @param req - The request sent by the client
- * @param res - The response to send to the client
- * @param next - The next function in the call chain
- * @returns The workflow UI page where the user can visualize the job as it happens
- */
-export async function getJobForWorkflowUI(
-  req: HarmonyRequest, res: Response, next: NextFunction,
-): Promise<void> {
-  const { jobID } = req.params;
-  try {
-    validateJobId(jobID);
-    const { page, limit } = getPagingParams(req, 1000);
-    const query: JobQuery = { requestId: jobID };
-    if (!req.context.isAdminAccess) {
-      query.username = req.user;
-    }
-    const { job } = await Job.byRequestId(db, jobID, 0, 0);
-    if (job) {
-      if (!(await job.canShareResultsWith(req.user, req.context.isAdminAccess, req.accessToken))) {
-        throw new NotFoundError();
-      }
-      res.render('jobs/workflow-job', {
-        job,
-        page,
-        limit,
-      });
-    } else {
-      throw new NotFoundError(`Unable to find job ${jobID}`);
-    }
-  } catch (e) {
-    req.context.logger.error(e);
-    next(e);
-  }
-}
-
-/**
- * Render the work items table for the workflow UI.
- *
- * @param req - The request sent by the client
- * @param res - The response to send to the client
- * @param next - The next function in the call chain
- * @returns The work items table HTML
- */
-export async function getWorkItemsForWorkflowUI(
-  req: HarmonyRequest, res: Response, next: NextFunction,
-): Promise<void> {
-  const { jobID } = req.params;
-  const { checkJobStatus } = req.query;
-  const badgeClasses = {};
-  badgeClasses[WorkItemStatus.READY] = 'primary';
-  badgeClasses[WorkItemStatus.CANCELED] = 'secondary';
-  badgeClasses[WorkItemStatus.FAILED] = 'danger';
-  badgeClasses[WorkItemStatus.SUCCESSFUL] = 'success';
-  badgeClasses[WorkItemStatus.RUNNING] = 'info';
-  try {
-    validateJobId(jobID);
-    const query: JobQuery = { requestId: jobID };
-    if (!req.context.isAdminAccess) {
-      query.username = req.user;
-    }
-    const { job } = await Job.byRequestId(db, jobID, 0, 0);
-    if (job) {
-      if (!(await job.canShareResultsWith(req.user, req.context.isAdminAccess, req.accessToken))) {
-        throw new NotFoundError();
-      }
-      if (([JobStatus.SUCCESSFUL, JobStatus.CANCELED, JobStatus.FAILED].indexOf(job.status) > -1) && checkJobStatus === 'true') {
-        // tell the client that the job has finished
-        res.status(204).json({ status: job.status });
-        return;
-      }
-      const { page, limit } = getPagingParams(req, env.defaultJobListPageSize);
-      const { workItems, pagination } = await getWorkItemsByJobId(db, job.jobID, page, limit, 'asc');
-      const pageLinks = getPagingLinks(req, pagination);
-      const nextPage = pageLinks.find((l) => l.rel === 'next');
-      const previousPage = pageLinks.find((l) => l.rel === 'prev');
-      setPagingHeaders(res, pagination);
-      res.render('jobs/workflow-items-table', {
-        workItems,
-        workflowItemUpdated() { return (new Date(this.updatedAt).toISOString()); },
-        workflowItemCreated() { return (new Date(this.createdAt).toISOString()); },
-        workflowItemBadge() { return badgeClasses[this.status]; },
-        workflowItemStep() { return sanitizeImage(this.serviceID); },
-        links: [
-          { ...previousPage, linkTitle: 'previous' },
-          { ...nextPage, linkTitle: 'next' },
-        ],
-        linkDisabled() { return (this.href ? '' : 'disabled'); },
-        linkHref() {
-          return (this.href ? this.href
-            .replace('/work-items', '')
-            .replace(/(&|\?)checkJobStatus=(true|false)/, '') : '');
-        },
-      });
-    } else {
-      throw new NotFoundError(`Unable to find job ${jobID}`);
-    }
-  } catch (e) {
-    req.context.logger.error(e);
-    next(e);
   }
 }
