+++ app/views/workflow-ui/job/index.mustache.html
@@ -14,7 +14,7 @@
         crossorigin="anonymous" defer></script>
     <script language="javascript" id="_fed_an_ua_tag"
         src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?agency=NASA&subagency=GSFC&dclink=true" defer></script>
-    <script language="javascript" type="module" src="../../../js/jobs/workflow-job.js" async></script>
+    <script language="javascript" type="module" src="../../../js/workflow-ui/job/index.js" async></script>
     <title>Harmony</title>
 </head>
 
@@ -35,7 +35,7 @@
         style="--bs-breadcrumb-divider: url(&#34;data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='8' height='8'%3E%3Cpath d='M2.5 0L1 1.5 3.5 4 1 6.5 2.5 8l4-4-4-4z' fill='currentColor'/%3E%3C/svg%3E&#34;);"
         aria-label="breadcrumb">
         <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a href="/workflow-ui/jobs">Jobs</a></li>
+            <li class="breadcrumb-item"><a href="/workflow-ui">Jobs</a></li>
             <li class="breadcrumb-item active" aria-current="page">{{job.jobID}}</li>
         </ol>
     </nav>
