+++ app/views/workflow-ui/jobs/index.mustache.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="../../../css/eui.min.css">
     <link rel="stylesheet" href="../../../css/default.css">
+    <script src="https://unpkg.com/vue@next"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
         integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
         crossorigin="anonymous" defer></script>
@@ -14,7 +15,7 @@
         crossorigin="anonymous" defer></script>
     <script language="javascript" id="_fed_an_ua_tag"
         src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?agency=NASA&subagency=GSFC&dclink=true" defer></script>
-    <script language="javascript" type="module" src="../../../js/jobs/workflow-jobs.js" defer></script>
+    <script language="javascript" type="module" src="../../../js/workflow-ui/jobs/index.js" defer></script>
     <title>Harmony</title>
 </head>
 
@@ -40,35 +41,7 @@
     </nav>
     <div class="container-lg">
         <div class="row pb-4">
-            <table class="table table-sm">
-                <thead>
-                    <tr>
-                        <th scope="col">jobID</th>
-                        <th scope="col">request</th>
-                        <th scope="col">status</th>
-                        <th scope="col">message</th>
-                        <th scope="col">numInputGranules</th>
-                        <th scope="col">progress</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {{#jobs}}
-                    <tr clas='job-table-row'>
-                        <th scope="row"><a href="jobs/{{jobID}}">{{jobID}}</a></th>
-                        <td class="mw-40ch">
-                            <small><code class="text-truncate d-block" title="{{request}}" data-bs-toggle="tooltip"
-                                data-bs-placement="right">{{jobUrl}}</code></small>
-                        </td>
-                        <td><span class="badge bg-{{jobBadge}}">{{status}}</span></td>
-                        <td class="mw-40ch"><small><code class="text-truncate d-block" title="{{message}}" data-bs-toggle="tooltip"
-                                    data-bs-placement="right">{{message}}</code></small></td>
-                        <td><span class="badge rounded-pill bg-warning text-dark">{{numInputGranules}}</span></td>
-                        <td>{{progress}}%</td>
-                    </tr>
-                    {{/jobs}}
-
-                </tbody>
-            </table>
+            {{> workflow-ui/jobs/jobs-table}}
             <nav aria-label="Page navigation" class="d-flex flex-column align-items-center">
                 <ul class="pagination px-0 mx-auto">
                     {{#links}}
