+++ test/helpers/jobs.ts
@@ -142,26 +142,6 @@ export function adminJobListing(app: Application, query: object = {}): Test {
 }
 
 /**
- * Makes a request to view the workflow UI jobs endpoint
- * @param app - The express application (typically this.frontend)
- * @param query - Mapping of query param names to values
- * @returns The response
- */
-export function workflowUIJobs(app: Application, query: object = {}): Test {
-  return request(app).get('/workflow-ui/jobs').query(query);
-}
-
-/**
- * Makes an admin request to view the workflow UI jobs endpoint
- * @param app - The express application (typically this.frontend)
- * @param query - Mapping of query param names to values
- * @returns The response
- */
-export function adminWorkflowUIJobs(app: Application, query: object = {}): Test {
-  return request(app).get('/admin/workflow-ui/jobs').query(query);
-}
-
-/**
  * Navigates to the job status route as the given user
  *
  * @param app - The express application (typically this.frontend)
@@ -184,40 +164,6 @@ export function adminJobStatus(app: Express.Application, options): Test {
   const { jobID, query } = options;
   const actualQuery = query || {};
   return request(app).get(`/admin/jobs/${jobID}`).query(actualQuery);
-}
-
-/**
- * Makes a request to view the workflow UI job endpoint
- *
- * @param app - The express application (typically this.frontend)
- * @param options - Mapping object. Includes (optional) query param object (which maps query
- * param names to values), jobID (to be used as the URL param), and (optional) username.
- * e.g. \{jobID: job.jobID, username: 'billy', query: \{...\}\}
- */
-export function workflowUIJob(
-  app: Express.Application,
-  options: { jobID: string; username?: string; query?: object },
-): Test {
-  const { jobID, query } = options;
-  const actualQuery = query || {};
-  return request(app).get(`/workflow-ui/jobs/${jobID}`).query(actualQuery);
-}
-
-/**
- * Makes an admin request to view the workflow UI job endpoint
- *
- * @param app - The express application (typically this.frontend)
- * @param options - Mapping object. Includes (optional) query param object (which maps query
- * param names to values), jobID (to be used as the URL param), and (optional) username.
- * e.g. \{jobID: job.jobID, username: 'billy', query: \{...\}\}
- */
-export function adminWorkflowUIJob(
-  app: Express.Application,
-  options: { jobID: string; username?: string; query?: object },
-): Test {
-  const { jobID, query } = options;
-  const actualQuery = query || {};
-  return request(app).get(`/admin/workflow-ui/jobs/${jobID}`).query(actualQuery);
 }
 
 /**
@@ -260,54 +206,14 @@ export function adminCancelJobWithGET(app: Express.Application, { jobID }: Job):
   return request(app).get(`/admin/jobs/${jobID}/cancel`);
 }
 
-/**
- * Makes a request to view the workflow UI work items endpoint
- *
- * @param app - The express application (typically this.frontend)
- * @param options - Mapping object. Includes (optional) query param object (which maps query
- * param names to values), jobID (to be used as the URL param), and (optional) username.
- * e.g. \{jobID: job.jobID, username: 'billy', query: \{...\}\}
- */
-export function workflowUIWorkItems(
-  app: Express.Application,
-  options: { jobID: string; username?: string; query?: object },
-): Test {
-  const { jobID, query } = options;
-  const actualQuery = query || {};
-  return request(app).get(`/workflow-ui/jobs/${jobID}/work-items`).query(actualQuery);
-}
-
-/**
- * Makes an admin request to view the workflow UI work items endpoint
- *
- * @param app - The express application (typically this.frontend)
- * @param options - Mapping object. Includes (optional) query param object (which maps query
- * param names to values), jobID (to be used as the URL param), and (optional) username.
- * e.g. \{jobID: job.jobID, username: 'billy', query: \{...\}\}
- */
-export function adminWorkflowUIWorkItems(
-  app: Express.Application,
-  options: { jobID: string; username?: string; query?: object },
-): Test {
-  const { jobID, query } = options;
-  const actualQuery = query || {};
-  return request(app).get(`/admin/workflow-ui/jobs/${jobID}/work-items`).query(actualQuery);
-}
-
 export const hookJobListing = hookRequest.bind(this, jobListing);
 export const hookAdminJobListing = hookRequest.bind(this, adminJobListing);
-export const hookWorkflowUIJobs = hookRequest.bind(this, workflowUIJobs);
-export const hookAdminWorkflowUIJobs = hookRequest.bind(this, adminWorkflowUIJobs);
 export const hookJobStatus = hookRequest.bind(this, jobStatus);
 export const hookAdminJobStatus = hookRequest.bind(this, adminJobStatus);
-export const hookWorkflowUIJob = hookRequest.bind(this, workflowUIJob);
-export const hookAdminWorkflowUIJob = hookRequest.bind(this, adminWorkflowUIJob);
 export const hookCancelJob = hookRequest.bind(this, cancelJob);
 export const hookAdminCancelJob = hookRequest.bind(this, adminCancelJob);
 export const hookCancelJobWithGET = hookRequest.bind(this, cancelJobWithGET);
 export const hookAdminCancelJobWithGET = hookRequest.bind(this, adminCancelJobWithGET);
-export const hookWorkflowUIWorkItems = hookRequest.bind(this, workflowUIWorkItems);
-export const hookAdminWorkflowUIWorkItems = hookRequest.bind(this, adminWorkflowUIWorkItems);
 
 /**
  * Given a string returns a new string with all characters escaped such that the string
